% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.seeds.R
\name{as.seeds}
\alias{as.seeds}
\title{format data as seeds}
\usage{
as.seeds(seeds = NULL, parents = NULL, generation = NULL)
}
\arguments{
\item{seeds}{a dataframe containing stage one seeds created directly from parents (default = NULL).}

\item{parents}{a dataframe containing all the viable parents on the landscape (default = NULL).}

\item{generation}{integer representing generation that seeds were produced in.}
}
\value{
a dataframe containing updated seeds.
}
\description{
Formats new seeds so they have all the right data regarding ploidy levels, generation and life_stage, while also removing seeds whose parents have ploidy levels that do not match.
}
\author{
Rose McKeon
}
