% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_combine.R
\name{count_combine}
\alias{count_combine}
\title{Combine counts from simulations into one data file.}
\usage{
count_combine(path_in, path_out)
}
\arguments{
\item{path_in}{character string contining the path, based on the current working directory, where the named subfolders that contain the simulation data are held.}

\item{path_out}{character string caontaining the path abd filename where the combined count file should be stored. File extension (.rds) added automatically.}
}
\description{
Takes all the simulation subfolders from a given data path_in and combines them into one rds file stored in path_out.
}
\examples{
count_combine("data/germination-benefit", "thesis/_data/germination_counts")
}
\author{
Rose McKeon
}
