% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicate_genomes.R
\name{duplicate_genomes}
\alias{duplicate_genomes}
\title{duplicate genomes}
\usage{
duplicate_genomes(pop = NULL, ploidy_rate = 0.01)
}
\arguments{
\item{pop}{A data frame containing individuals with a numerical column 'ploidy' that will be doubled (default = NULL).}

\item{ploidy_rate}{A number between 0 and 1 which represents the rate at which genome duplication occurs (default = 0.01).}
}
\value{
a dataframe containing the updated population.
}
\description{
Double the ploidy value of random individuals within a population data frame at a specified rate.
}
\author{
Rose McKeon
}
