% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_log.R
\name{setup_log}
\alias{setup_log}
\title{Log output to `tempdir()`.}
\usage{
setup_log()
}
\value{
character string containing the location of the temporary file.
}
\description{
Creates temporary text file containing simulation output.
}
\examples{
setup_log()
}
\author{
Rose McKeon
}
