% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_temp_files.R
\name{setup_tmp_files}
\alias{setup_tmp_files}
\title{Ensure temp files are ready for simulation data}
\usage{
setup_temp_files()
}
\arguments{
\item{generation}{integer representing the generation that this data represents}
}
\description{
Creates temporary rds files for storing generation information separated by life stage and conditionally creates log files in text format.
}
\examples{
setup_tmp_files(seedbank, juveniles, adults, seedoutput, 1)
}
\author{
Rose McKeon
}
