% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sploidy.R
\name{sploidy}
\alias{sploidy}
\title{What are the relative densities of polyploids vs diploids?}
\usage{
sploidy()
}
\arguments{
\item{pop_size}{numerical vector of integers representing starting population size by life stage (default = c(5, 5, 5)).}

\item{grid_size}{integer representing the size of the landscape grid. Cells are numbered 0 to grid_size -1 along an X and Y axis (default = 10, so the grid is 10 x 10).}

\item{ploidy_rate}{number between 0 and 1 representing the chance that genome duplication will occur (default = 0, so no genome duplication).}

\item{trans}{matrix (3x3) representing the tranisiton probabilities between life stages 1 to 3. Values [1,1] and [2,1] are not used. These transtitons are a product of G*D instead (default = NULL).}

\item{G}{numerical value representing the rate of germination (default = 0.5).}

\item{G_modifier}{numerical value representing the factor by which to multiply G for polyploid seeds (default = 1, no change).}

\item{D}{numerical value representing the seed survival rate (default = 0.5).}

\item{K}{numerical value representing the carrying capacity of the entire landscape (default = grid_size x grid_size x 10).}

\item{seed_longevity}{integer representing the maximum seed life during dormancy (default = 1).}

\item{generations}{integer representing the number of generations the model should attempt to run for (default = 10). The simulation will break early if extinction occurs.}

\item{simulations}{integer representing the number of simulations which should be run with these parameters (default = 2).}

\item{filepath}{character string defining the file path where output files should be stored (relative to working directory). If this folder does not exist it will be created (default = "data").}

\item{name}{character string defining the name of the filepath subfolder which countains output RDS files. (default = unique random identifier).}

\item{log}{logical value which indicates whether a log file with verbose output messages should be stored in filepath/name (default = T).}

\item{output_gen_data}{logical value that defines whether population data contining info on all individuals at the end of every generation is stored (default = F, storage intensive).}
}
\value{
Nothing is returned. Messages are output to the console so you can see each generation complete, and data output is stored according to the filepath and name paremeters.
}
\description{
A spatially explicit individual-based model which runs a simulation, or repeated simulations, of a plant population over time.
}
\author{
Rose McKeon
}
