% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stop_log.R
\name{stop_log}
\alias{stop_log}
\title{Stop logging messages and clear log files from `tempdir()`.}
\usage{
stop_log()
}
\value{
character string containing the location of the temporary file.
}
\description{
Closes the connection to log file passed form log_info$connection and unlinks the log file passed from log_info$path.
}
\examples{
log_info <- setup_log()
stop_log(log_info)
}
\author{
Rose McKeon
}
