% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store_session.R
\name{store_session}
\alias{store_session}
\title{Store session info}
\usage{
store_session()
}
\arguments{
\item{sploidy_call}{mat.call object passed from main sploidy function.}

\item{name}{character string used in folder structure.}
}
\description{
Creates a file containing timestamp simulation was begun, the function call it ran with and the session info with package version info for all the loaded libraries.
}
\examples{
store_session(match.call(), "whatever")
}
\author{
Rose McKeon
}
