% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/store_tmp_data.R
\name{store_tmp_data}
\alias{store_tmp_data}
\title{Store object as RDS file in temp directory.}
\usage{
store_tmp_data()
}
\arguments{
\item{object}{any R data object to store.}

\item{tmp_filename}{character string used in file naming.}
}
\value{
character string containing the location of the temporary file.
}
\description{
Creates temporary rds file from object.
}
\examples{
seedbank <- "whatever"
seedbank_location <- store_tmp_data(seedbank, "seedbank")
}
\author{
Rose McKeon
}
