% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survive.R
\name{survive}
\alias{survive}
\title{survive}
\usage{
survive(pop, prob = 1, dead = F)
}
\arguments{
\item{pop}{a population dataframe as generated by populate_landscape(), ie: nested by plant.}

\item{prob}{a number between 0 and 1 representing probability of survival (default = 1, all are very likely to survive).}

\item{dead}{logical value that determines if the dead are returned (default = FALSE).}
}
\value{
dataframe of survivors, or a list containing survivors as well as the dead.
}
\description{
Reduces population data based on probability of survival, or alternatively also returns the dead.
}
\author{
Rose McKeon
}
